style: Cleanup duplicated code in the definition of ComputedValues.
